Sorting out some issues with MediaAttachmentType enum.

+++ victorious/VictoriousIOSSDK/FollowCountRequest.swift
@@ -21,8 +21,8 @@ public struct FollowCountRequest: RequestType {
     }
     
     public func parseResponse(response: NSURLResponse, toRequest request: NSURLRequest, responseData: NSData, responseJSON: JSON) throws -> FollowCount {
-        guard let followingCount = Int(responseJSON["payload"]["subscribed_to"].stringValue),
-            let followersCount = Int(responseJSON["payload"]["followers"].stringValue) else {
+        guard let followingCount = Int64(responseJSON["payload"]["subscribed_to"].stringValue),
+            let followersCount = Int64(responseJSON["payload"]["followers"].stringValue) else {
                 throw ResponseParsingError()
         }
         return FollowCount(followingCount: followingCount, followersCount: followersCount)
@@ -30,6 +30,6 @@ public struct FollowCountRequest: RequestType {
 }
 
 public struct FollowCount {
-    public let followingCount: Int
-    public let followersCount: Int
+    public let followingCount: Int64
+    public let followersCount: Int64
 }
